requirements.txt
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
python-socketio==5.10.0
python-engineio==4.8.0
bcrypt==4.1.1
python-dotenv==1.0.0
eventlet==0.33.3
